backend/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==2.5.0
greenlet==3.1.1
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==3.20.1
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.0.0
redis==5.0.0
SQLAlchemy==1.4.39
structlog==23.1.0
typing_extensions==4.12.2
Werkzeug==3.1.3
